@@ -1,7 +1,6 @@
 import asyncio
-from abc import ABC, abstractmethod
 from time import time
-from typing import TYPE_CHECKING, Any, Coroutine, Generic, Optional, Union
+from typing import TYPE_CHECKING, Any, Generic, Optional
 
 from typing_extensions import TypeVar
 
@@ -19,70 +18,7 @@ if TYPE_CHECKING:  # pragma: no cover
 _ReturnType = TypeVar("_ReturnType")
 
 
-class _Task(ABC, Generic[_ReturnType]):
-    """TaskiqTask interface."""
-
-    @abstractmethod
-    def is_ready(self) -> Union[bool, Coroutine[Any, Any, bool]]:
-        """
-        Method to check wether result is ready.
-
-        :return: True if result is ready.
-        """
-
-    @abstractmethod
-    def get_result(
-        self,
-        with_logs: bool = False,
-    ) -> Union[
-        "TaskiqResult[_ReturnType]",
-        Coroutine[Any, Any, "TaskiqResult[_ReturnType]"],
-    ]:
-        """
-        Get actual execution result.
-
-        :param with_logs: wether you want to fetch logs.
-        :return: TaskiqResult.
-        """
-
-    @abstractmethod
-    def wait_result(
-        self,
-        check_interval: float = 0.2,
-        timeout: float = -1.0,
-        with_logs: bool = False,
-    ) -> Union[
-        "TaskiqResult[_ReturnType]",
-        Coroutine[Any, Any, "TaskiqResult[_ReturnType]"],
-    ]:
-        """
-        Wait for result to become ready and get it.
-
-        This function constantly checks whether result is ready
-        and fetches it when it becomes available.
-
-        :param check_interval: how often availability is checked.
-        :param timeout: maximum amount of time it will wait
-            before raising TaskiqResultTimeoutError.
-        :param with_logs: whether you need to download logs.
-        :return: TaskiqResult.
-        """
-
-    @abstractmethod
-    def get_progress(
-        self,
-    ) -> Union[
-        "Optional[TaskProgress[Any]]",
-        Coroutine[Any, Any, "Optional[TaskProgress[Any]]"],
-    ]:
-        """
-        Get task progress.
-
-        :return: task's progress.
-        """
-
-
-class AsyncTaskiqTask(_Task[_ReturnType]):
+class AsyncTaskiqTask(Generic[_ReturnType]):
     """AsyncTask for AsyncResultBackend."""
 
     def __init__(
